requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

torch_geometric>=2.3.0
pyg_lib
torch_scatter
torch_sparse
torch_cluster
torch_spline_conv

pytorch-lightning>=2.0.0

 
easydict>=1.10
type_enforced>=0.0.4
pyyaml>=6.0

numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0

wandb>=0.15.0
tensorboard>=2.13.0

jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.